xmlui: Switch all XMLUI themes to non-caching Cocoon pipelines
We seem to be having overly-excessive caching on XMLUI and it's
causing our themes to show up incorrectly.  I'm not sure of the
performance implications of this for the server, but it seemed
to work well for the CCAFS theme initially and the issues are
persisting with other themes now so we might as well try.

Possibly related: https://jira.duraspace.org/browse/DS-298