DS-2027 Update bitstream-ordering.js
https://jira.duraspace.org/browse/DS-2027

There is a bad JQuery selector in the following code. This causes a submit to take place each time an up or down arrow is clicked. It is difficult to see the error in the javascript console because the submit takes place immediately.

Due to the page refresh delay, this is confusing to users when a large number of bitstreams are present.

Note that "$^=" is not a valid selector.

http://api.jquery.com/category/selectors/