Statistics.xml: Remove country-specific stats from item view
We have to remove this now because the geo data recorded for item
accesses for the last three years has been mis-configured, and was
logging the local host's IP rather than the request originator's IP.

This is due to using Apache mod_proxy / nginx to sit in front of
Tomcat, without turning on DSpace's `useProxies` option, which would
have recorded the X-Forwarded-* headers instead.

We can perhaps re-enable these stats in a few months when we actually
have real data, and limit it to the past 6 months or something.

Signed-off-by: Alan Orth <a.orth@cgiar.org>