messages.xml: Add strings for CIP subjects
Addresses #46.

Still to figure out which strings we need to print as "subject" and
which we need to print as "subjects".  From looking at the XMLUI
string names it's not exactly obvious where they will be used, so it
makes it hard to judge what context they will be printed in!

Signed-off-by: Alan Orth <a.orth@cgiar.org>