dspace.cfg: Fixes for Atmire's Listings and Reports
Somehow this slipped through during the DSpace 3.x -> 4.x rebase.
Worse, Listings and Reports uses Lucene (and everything else uses
Solr), and we haven't updated the indexes since the move to 4.x,
which means the module has been working with old data in CGSpace
since December, 2014, which means we probably need to do an Lucene
init for good measure, and add a cron job for the Lucene index
update periodically, perhaps with the same frequ-
ency of `index-discovery`.