Clean up dc.type filters in Discovery and Atmire CUA
We have two separate filters for the same fields, and we use them
in two different places. One is dc.type, which is a vanilla Dublin
Core value, and the other is dc.type.output which we somehow invented
a few years ago. We can use one search filter to catch both of these
and then we can get rid of the other one.

Requires a full Discovery re-index (dspace index-discovery -b).

Signed-off-by: Alan Orth <alan.orth@gmail.com>