Update Atmire Listings and Reports for Output→Type change
This was forgotten during the move last week and broke the Listings
and Reports module. I'm not sure why we don't use `dc.type.*` for
the schema and something other than `type` for the Discovery field,
but I will have to ask Atmire people about it because the docs are
not very specific, and this works.

Signed-off-by: Alan Orth <alan.orth@gmail.com>