dspace.cfg: Use X-Forwarded-For header in admin activity menu
I'm not sure why, but this is actually already working for some
reason. We have the following in our front-end nginx config:

    proxy_set_header X-Real-IP $remote_addr;
    proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;

... so maybe DSpace is checking X-Real-IP instead, or checking
X-Forwarded-For already despite this config. In any case, it's
better to have this enabled to be consistent with stated intent,
just in case this is a bug or something that gets fixed later and
breaks because we were relying on it!