Add index for CGIAR System subject
When we migrated the CGIAR Library to CGSpace in 2017-09 we moved their
dc.subject metadata to cg.subject.system. Their website has many links
to browse and search results for these subjects so we need to accomodate
them somehow. I will capture browse requests for "type=subject" in nginx
and redirect them to the new index (systemsubject).