Import atmire-cua-scheduled-tasks.xml from module
This tasks appears to have a bug that causes it to create a very high
load on the server, and I think this might be the reason that CGSpace
crashes every morning. I am importing the config from the module so I
can adjust the timing of the taskUpdateSolrStatsMetadata task.