dspace/config: Adjust timing of taskUpdateSolrStatsMetadata task
I suspect this task is responsible for creating extremely high load on
the server every morning. Whenever CGSpace goes down in the morning, I
see this script has been running for several hours, there are hundreds
of database connections, and the nginx access logs don't show anything
out of the ordinary. I will move this job from 3AM to 3PM to see if it
changes anything.