xmlui: Enable display of author ORCID iDs
If the cg.creator.id field is present it will be displayed along with a
link to the identifier's ORCID profile. The display is according to the
ORCID guidelines. I settled on using a badge size of 1em because larger
sizes like Font Awesome's 2x or even simply setting 1.5em in CSS cause
the text's line height to be too large and awkward.

See: https://orcid.org/trademark-and-id-display-guidelines